cáestudy

+++ app/Http/Controllers/CategoryController.php
@@ -15,6 +15,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+      $this->authorize('viewAny', Category::class);
       $categories = Category::paginate(3);
       $param =[
         'categories'=> $categories
@@ -29,6 +30,7 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Category::class);
         return view('category.add');
     }
 
@@ -79,6 +81,7 @@ class CategoryController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', Category::class);
         $categories = Category::find($id);
         return view('category.edit', compact(['categories']));
     }
@@ -108,6 +111,7 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('forceDelete', Category::class);
         $category=Category::onlyTrashed()->findOrFail($id);
         $category->forceDelete()->with('status','Xóa danh mục thành công');
 
@@ -122,6 +126,7 @@ class CategoryController extends Controller
     return view('category.index', compact('categories'));
     }
     public  function softdeletes($id){
+        $this->authorize('delete', Category::class);
         date_default_timezone_set("Asia/Ho_Chi_Minh");
         $category = Category::findOrFail($id);
         $category->deleted_at = date("Y-m-d h:i:s");
@@ -129,11 +134,13 @@ class CategoryController extends Controller
         return redirect()->route('category.index');
     }
     public  function trash(){
+        $this->authorize('viewtrash', Category::class);
         $categories = Category::onlyTrashed()->get();
         $param = ['categories'    => $categories];
         return view('category.trash', $param);
     }
     public function restoredelete($id){
+        $this->authorize('restore', Category::class);
         $categories=Category::withTrashed()->where('id', $id);
         $categories->restore();
         return redirect()->route('category.trash');
